# docs/openapi3.yaml
openapi: 3.0.0
info:
  description: Production-ready backend Endge-сервиса с эталонными
    RedPanda-обёртками и строгими архитектурными границами.
  title: Endge Service Backend API
  contact: {}
  version: 1.0.20
paths:
...
  "/api/v1/projects/{project_identity}/queries":
    get:
      security:
        - BearerAuth: []
      description: Возвращает активные Query проекта. Можно отфильтровать записи по
        identity папки и query type; soft-deleted Query не возвращаются.
      tags:
        - queries
      summary: Список Query
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: root-queries
          description: Folder identity
          name: folder_identity
          in: query
          schema:
            type: string
        - example: http
          description: Query type
          name: query_type
          in: query
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_query.QueriesListResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    post:
      security:
        - BearerAuth: []
      description: Создает конфигурацию Query в папке проекта. Query source, headers,
        параметры и mock data только хранятся и не исполняются сервисом.
      tags:
        - queries
      summary: Создать Query
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_query.CreateQueryRequest"
        description: Параметры Query
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_query.QueryResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/queries/{query_identity}":
    get:
      security:
        - BearerAuth: []
      description: Возвращает активную Query по identity в пределах проекта вместе с
        identity ее папки.
      tags:
        - queries
      summary: Получить Query
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: users-list
          description: Query identity
          name: query_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_query.QueryResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    delete:
      security:
        - BearerAuth: []
      description: Выполняет soft-delete Query. Связанные DataView физически не
        удаляются, но не попадают в обычные query-based сценарии.
      tags:
        - queries
      summary: Удалить Query
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: Query identity
          name: query_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
    patch:
      security:
        - BearerAuth: []
      description: Полностью заменяет editable payload Query, включая папку, source,
        параметры, headers, auth и mock data. Поля id, identity, createdAt и
        deletedAt сохраняются.
      tags:
        - queries
      summary: Обновить Query
      parameters:
        - example: demo-project
          description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - example: users-list
          description: Query identity
          name: query_identity
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_api_http_v1_query.UpdateQueryRequest"
        description: Параметры обновления Query
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/internal_api_http_v1_query.QueryResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/queries/{query_identity}/hard":
    delete:
      security:
        - BearerAuth: []
      description: Выполняет hard-delete soft-deleted Query. Связанные DataView
        удаляются каскадно на уровне базы данных.
      tags:
        - queries
      summary: Физически удалить Query
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: Query identity
          name: query_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
  "/api/v1/projects/{project_identity}/queries/{query_identity}/restore":
    post:
      security:
        - BearerAuth: []
      description: Восстанавливает soft-deleted Query по identity в пределах проекта.
      tags:
        - queries
      summary: Восстановить Query
      parameters:
        - description: Project identity
          name: project_identity
          in: path
          required: true
          schema:
            type: string
        - description: Query identity
          name: query_identity
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "404":
          description: Not Found
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
        "500":
          description: Internal Server Error
          content:
            "*/*":
              schema:
                $ref: "#/components/schemas/github_com_endge-lab_service-backend_internal_api_h\
                  ttp_v1_transport.ErrorResponse"
...
servers:
  - url: /
components:
  securitySchemes:
    BearerAuth:
      type: apiKey
      name: Authorization
      in: header
  schemas:
...
    internal_api_http_v1_query.CreateQueryRequest:
      type: object
      required:
        - displayName
        - folderIdentity
        - identity
        - queryType
        - source
      properties:
        active:
          type: boolean
        auth:
          type: object
          additionalProperties: {}
        description:
          type: string
        displayName:
          type: string
        folderIdentity:
          type: string
        headers:
          type: object
          additionalProperties: {}
        identity:
          type: string
        meta:
          type: object
          additionalProperties: {}
        mockData:
          type: object
          additionalProperties: {}
        mockDataEnabled:
          type: boolean
        params:
          type: array
          items: {}
        queryType:
          type: string
        source:
          type: object
          additionalProperties: {}
        timeoutMs:
          type: integer
    internal_api_http_v1_query.QueriesListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/internal_api_http_v1_query.QueryResponse"
    internal_api_http_v1_query.QueryResponse:
      type: object
      properties:
        active:
          type: boolean
        auth:
          type: object
          additionalProperties: {}
        createdAt:
          type: string
        deletedAt:
          type: string
        description:
          type: string
        displayName:
          type: string
        folderIdentity:
          type: string
        headers:
          type: object
          additionalProperties: {}
        id:
          type: string
        identity:
          type: string
        meta:
          type: object
          additionalProperties: {}
        mockData:
          type: object
          additionalProperties: {}
        mockDataEnabled:
          type: boolean
        params:
          type: array
          items: {}
        projectIdentity:
          type: string
        queryType:
          type: string
        source:
          type: object
          additionalProperties: {}
        timeoutMs:
          type: integer
        updatedAt:
          type: string
    internal_api_http_v1_query.UpdateQueryRequest:
      type: object
      required:
        - displayName
        - folderIdentity
        - queryType
        - source
      properties:
        active:
          type: boolean
        auth:
          type: object
          additionalProperties: {}
        description:
          type: string
        displayName:
          type: string
        folderIdentity:
          type: string
        headers:
          type: object
          additionalProperties: {}
        meta:
          type: object
          additionalProperties: {}
        mockData:
          type: object
          additionalProperties: {}
        mockDataEnabled:
          type: boolean
        params:
          type: array
          items: {}
        queryType:
          type: string
        source:
          type: object
          additionalProperties: {}
        timeoutMs:
          type: integer
...